Write createUser static
All of the user-creation logic goes in the createUser static, not in any
controller.

I've never used sinon before, so the tests for this static might be a
bit crude (and sinon's awesome once you know what it does)

I'm not sure how to test the pre-save hook though, at least not without
connecting to a database.